refactor[app]: file structure
