--- client/src/App.css
@@ -1,47 +1,0 @@
-.app {
-  --app-header-height: 64px;
-}
-
-.app-content .display {
-  min-height: calc(100vh - var(--app-header-height));
-  padding: 20px;
-}
-
-.app-content .display [id] {
-  scroll-margin-top: var(--app-header-height);
-}
-
-.app-footer {
-  text-align: center;
-}
-
-.app-footer .social-media {
-  margin-top: 12px;
-}
-
-.app-footer .footer-link {
-  color: black !important;
-}
-
-.app-footer .footer-icon {
-  font-size: 24px;
-}
-
-.cherry-markdown {
-  word-break: normal !important;
-}
-
-#cherry-editor .cherry-dropdown[name="insert"] {
-  width: 180px;
-}
-
-#cherry-viewer .cherry {
-  box-shadow: none !important;
-  background: transparent !important;
-}
-
-#cherry-viewer .cherry-previewer {
-  border: none !important;
-  background: transparent !important;
-  padding: 0 16px !important;
-}
