+++ client/src/modules/blog/components/CherryEditor.css
@@ -1,0 +1,18 @@
+.cherry-markdown {
+  word-break: normal !important;
+}
+
+#cherry-editor .cherry-dropdown[name="insert"] {
+  width: 180px;
+}
+
+#cherry-viewer .cherry {
+  box-shadow: none !important;
+  background: transparent !important;
+}
+
+#cherry-viewer .cherry-previewer {
+  border: none !important;
+  background: transparent !important;
+  padding: 0 16px !important;
+}
